Force indices to be long datatype

+++ brainspy/processors/dnpu.py
@@ -195,8 +195,9 @@ class DNPU(nn.Module):
         data_input_indices : Sequence[int]
             Indices of the input electrodes.
         """
-        self.register_buffer("data_input_indices",
-                             torch.tensor(data_input_indices))
+        self.register_buffer(
+            "data_input_indices",
+            torch.tensor(data_input_indices, dtype=torch.long))
         self.node_no = self.init_node_no()
         self.data_input_electrode_no, self.control_electrode_no = self.init_activation_electrode_no(
         )
@@ -231,8 +232,9 @@ class DNPU(nn.Module):
             "control_ranges",
             torch.stack([voltage_ranges[i] for i in control_list]))
 
-        self.register_buffer("control_indices",
-                             torch.tensor(np.array(control_list)))
+        self.register_buffer(
+            "control_indices",
+            torch.tensor(np.array(control_list, dtype=torch.long)))
         # # Define control electrode indices
         # self.control_indices = TorchUtils.format(
         #     control_list, data_type=torch.int64
